feat: remove Line Notify support, Slack-only notification
