+++ src/index.ts
@@ -25,41 +25,27 @@ export interface ISlackEventNotify {
  */
 export interface EventNotifyProps {
   /**
-   * Line Notify Token for Lambda send notify permission.
-   *
-   * @default - none
-   */
-  readonly lineNotifyToken?: string | undefined;
-
-  /**
    * Notify target to Slack channel.
-   *
-   * @default - none
    */
-  readonly slack?: ISlackEventNotify;
+  readonly slack: ISlackEventNotify;
 }
 
 /**
  * Event Notfiy Construct Class.
  */
 export class EventNotify extends Construct {
-  constructor(scope: Construct, id: string, props?: EventNotifyProps) {
+  constructor(scope: Construct, id: string, props: EventNotifyProps) {
     super(scope, id);
 
-    if (!props?.lineNotifyToken && !props?.slack) {
-      throw new Error('Please input lineNotifyToken or slack options');
-    }
-
     const lambdaFun = new lambda.Function(this, 'lambda_fun', {
-      handler: 'linenotify.lambda_handler',
+      handler: 'notify.lambda_handler',
       code: lambda.Code.fromAsset(path.join(__dirname, '../function') ),
       runtime: lambda.Runtime.PYTHON_3_12,
       timeout: cdk.Duration.minutes(3),
       logRetention: logs.RetentionDays.THREE_DAYS,
       environment: {
-        LINE_NOTIFY_TOKEN: props?.lineNotifyToken ? props?.lineNotifyToken : 'none',
-        SLACK_WEBHOOK_URL: props?.slack?.slackWebhookUrl ? props?.slack?.slackWebhookUrl : 'none',
-        SLACK_CHANNEL_NAME: props?.slack?.slackChannelName ? props?.slack?.slackChannelName : 'none',
+        SLACK_WEBHOOK_URL: props.slack.slackWebhookUrl,
+        SLACK_CHANNEL_NAME: props.slack.slackChannelName,
       },
     });
